+++ lib/importers.rb
@@ -1,0 +1,1 @@
+Dir[File.dirname(__FILE__) + '/importers/*.rb'].each {|file| require file }
