+++ lib/importers/scenario_importer.rb
@@ -2,20 +2,15 @@ require 'csv'
 
 class ScenarioImporter
 
-  def initialize(scenarios)
-    @scenarios = scenarios
-  end
-
-  def self.from_dir(dir)
-    scenarios = []
+  def initialize(dir)
+    @scenarios = []
     Dir.foreach(dir) do |name|
       path = File.join(dir,name)
-      scenarios << self.from_file(path) unless File.directory?(path)
+      @scenarios << self.from_file(path) unless File.directory?(path)
     end
-    ScenarioImporter.new(scenarios)
   end
 
-  def save(profile_type)
+  def import(profile_type)
     ActiveRecord::Base.transaction do
       context = Context.new do |c|
         c.context_type = "tests"
